pyproject.toml
[tool.poetry]
name = "my-minecraft-server"
version = "0.1.0"
description = "Minecraft 伺服器管理機器人"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "src" },
    { include = "scripts" },
    { include = "assets" }
]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
discord-py = "^2.3.2"
python-dotenv = "^1.0.0"
plyer = "^2.1.0"
pynacl = "^1.5.0"
psutil = "^6.1.1"
python-frontmatter = "^1.1.0"
discord = "^2.3.2"
setuptools = "^75.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mcbot = "scripts.run:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
